When forecasting, assume that all Heavy RIs will be fully utilized
Heavy RIs have all of their usage billed up front, and only over time is
this usage then allocated to specific instances.  Ideally we would
forecast based on usage, to correctly forecast charges for underutilized
heavy RIs, but there doesn't seem to be enough data to do that in the
cost allocation report.